+++ src/fhir/STU3R4-common/sdc-export.js
@@ -63,6 +63,13 @@ function addSTU3R4ExportFns(ns) {
         var itemValue = values[i];
         if(itemValue !== undefined && itemValue !== null && itemValue !== '') {
           var answer = null;
+          var ext = [];
+          if (itemValue.score !== null && itemValue.score !== undefined) {
+            ext.push({
+              url: this.fhirExtUrlOptionScore,
+              valueDecimal: itemValue.score,
+            });
+          }
           // with an answer list of Coding
           if (dataType === 'CODING') {
             // for optionsOrString, the value could be string if it is a user typed, not-on-list value
@@ -73,6 +80,9 @@ function addSTU3R4ExportFns(ns) {
               var answerCoding = this._setIfHasValue(null, 'system', LForms.Util.getCodeSystem(itemValue.system));
               answerCoding = this._setIfHasValue(answerCoding, 'code', itemValue.code);
               answerCoding = this._setIfHasValue(answerCoding, 'display', itemValue.text);
+              if (answerCoding && ext.length > 0) {
+                answerCoding.extension = ext;
+              }
               answer = this._setIfHasValue(null, 'valueCoding', answerCoding);
             }
           }
@@ -81,6 +91,9 @@ function addSTU3R4ExportFns(ns) {
             if (typeof itemValue !== 'string') { //R4 does not allow off list string
               var valueKey = this._getValueKeyByDataType("value", item);
               answer = {[valueKey]: itemValue.text};
+              if (ext.length > 0) {
+                answer['_'+valueKey] = {extension: ext};
+              }
             }
           }
           // without an answer list
@@ -95,12 +108,19 @@ function addSTU3R4ExportFns(ns) {
             //   "system" : "<uri>", // Code System that defines coded unit form
             //   "code" : "<code>" // Coded form of the unit
             // }]
-            answer = this._setIfHasValue(null, 'valueQuantity', this._makeValueQuantity(itemValue, item.unit));
+            var answerQuantity = this._makeValueQuantity(itemValue, item.unit);
+            if (answerQuantity && ext.length > 0) {
+              answerQuantity.extension = ext;
+            }
+            answer = this._setIfHasValue(null, 'valueQuantity', answerQuantity);
           }
           // for boolean, decimal, integer, date, dateTime, instant, time, string, uri, attachment, (no coding)
           else if (this._lformsTypesToFHIRFields[dataType]) {
             var valueKey = this._getValueKeyByDataType("value", item);
             answer = {[valueKey]: itemValue};
+            if (ext.length > 0) {
+              answer['_'+valueKey] = {extension: ext};
+            }
           }
         }