+++ CHANGELOG.md
@@ -2,9 +2,14 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
-## [36.4.0] 2024-07-22
+## [36.4.0] 2024-07-23
 ### Added
 - Support for rendering-xhtml on item.text.
+
+## [36.3.3] 2024-07-22
+### Added
+- Copy questionnaire-ordinalValue (STU3), ordinalValue (R4) and
+  itemWeight (R5) extensions to questionnaire response.
 
 ## [36.3.2] 2024-07-10
 ### Changed